VirtIO: 64bits features & create virtqueue ops
Most VirtIO implementation use 64bits values for features.
It's enough for most if not all VirtIO drivers.
32bits is not enough for newer drivers.

We need to use these new create virtqueue and delete virtqueue...
this enable some drivers to handle differently the virtqueues creation.
For a device driver it seams necessary to add new ops,
since they are not creating the virqueues directly.

+++ lib/virtio_mmio/virtio_mmio.c
@@ -20,11 +20,11 @@ extern uint32_t virtio_get_devid(const struct virtio_device *vdev);
 extern uint8_t virtio_device_get_status(struct virtio_device *vdev);
 extern void virtio_device_set_status(struct virtio_device *vdev, uint8_t status);
 extern void virtio_device_read_config(struct virtio_device *vdev,
-			       uint32_t offset, void *dst, int length);
+				      uint32_t offset, void *dst, int length);
 extern void virtio_device_write_config(struct virtio_device *vdev,
-				uint32_t offset, void *src, int length);
-extern uint32_t virtio_device_get_features(struct virtio_device *vdev);
-extern void virtio_device_set_features(struct virtio_device *vdev, uint32_t features);
+				       uint32_t offset, void *src, int length);
+extern uint64_t virtio_device_get_features(struct virtio_device *vdev);
+extern void virtio_device_set_features(struct virtio_device *vdev, uint64_t features);
 extern void virtio_device_reset(struct virtio_device *vdev);
 
 void virtio_mmio_isr(struct virtio_device *vdev);
